HOTFIX: Aaron 01 24 24 email grader improvements (#1798)
* Fixed bug

* Added valid to other dicts

* Used lemmatization and regex to improve spam detection

* Updated body math

* Fixed bug

* Added feedback point

* Fixed casting

---------

+++ src/email_sequencing/services.py
@@ -745,7 +745,7 @@ def grade_email(tracking_data: dict, subject: str, body: str):
     goods += feedback_point
 
     total_checks = 6
-    feedback_score = (goods / total_checks) * 100
+    feedback_score = int((goods / total_checks) * 100)
 
     # Create a record in the database
     entry = EmailGraderEntry(
